Added Affiliate Program sectyion
Mystery: LIne 97 follows the general/faqs sectrion and others in listing the index file first within the category ("general/affiliate-program", // Index). However, the affiliate-program.mdx file appears as the first item in the sidebar under "Affiliiate Programs", wheras the other categories with the index listed first in sidebar.js don't show this fle in the sidebar.  I can't find how to keep it from appearing. If I remove it, everything seems to work well but it's not consistent with other categories.